% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai_from_ao.R
\name{ai_from_ao}
\alias{ai_from_ao}
\title{Solve for Inherent Availability Given Operational Availability, CMT and ALDT}
\usage{
ai_from_ao(ao, cmt, aldt)
}
\arguments{
\item{ao}{Operational Availability.}

\item{cmt}{Corrective Maintenance Time.}

\item{aldt}{Administrative and Logistics Down Time.}
}
\value{
The output will be a numeric value for Inherent Availability.
}
\description{
\code{ai_from_ao} solves for Inherent Availability
  given the other inputs.
}
\examples{
# What would inherent availability be if operational availability was 0.83,
 # CMT was 10.8 hours, and ALDT was 40.4 hours?
ai_from_ao(ao = 0.83, cmt = 10.8, aldt = 40.4)

}
\seealso{
\code{\link{ao_unknown}}, \code{\link{ao_keesee}},
  \code{\link{ao_keesee_test_duration}}
}
