% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amsaa.R
\docType{data}
\name{amsaa}
\alias{amsaa}
\title{Repairable systems data}
\format{
A data frame with 36 rows and 3 variables:
\describe{
  \item{Failure}{Failure numbers for each sytem.}
  \item{System}{System identifier, S1, S2, S3.}
  \item{Time}{Time in hours at which failures occurred.}
}
}
\usage{
amsaa
}
\description{
Simlated data from AMSAA technical report 138. This contains simulated
data consisting of the failure times for three repairable systems.
}
\keyword{datasets}
