% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ao_keesee.R
\name{ao_keesee}
\alias{ao_keesee}
\title{CI for Operational Availability as given by Keesee}
\usage{
ao_keesee(n, m, upt, dnt, alpha = 0.4)
}
\arguments{
\item{n}{Number of failures.}

\item{m}{Number of repairs.}

\item{upt}{Up Time.}

\item{dnt}{Down Time.}

\item{alpha}{Two-sided alpha value.}
}
\value{
The output will be a numeric vector giving the two-sided CI for
  Operational Availability, and point estimate.
}
\description{
\code{ao_keesee} gives a 2-sided CI for Operational Availability.
  The method was documented by W.R. Keesee in "A Method of Determining
  a Confidence Interval for Availability".
}
\examples{
# What is the 90\% 2-sided Ao CI given 14 failures, 20 repairs,
  # 200 hours of Up Time and 100 hours of Down Time?
ao_keesee(n = 14, m = 20, upt = 200, dnt = 100, alpha = .2)

}
\references{
Keesee, W.R. A Method of Determining a Confidence Interval for Availability. 1965. url{https://apps.dtic.mil/sti/citations/AD0617716}
}
\seealso{
\code{\link{ao_unknown}}, \code{\link{ai_from_ao}},
  \code{\link{ao_keesee_test_duration}}
}
