% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ao_keesee_test_duration.R
\name{ao_keesee_test_duration}
\alias{ao_keesee_test_duration}
\title{Determines the test time required to demonstrate Availability was met with Confidence}
\usage{
ao_keesee_test_duration(n, m, k = NA, mttr, mldt = NA, ao, alpha = 0.4)
}
\arguments{
\item{n}{Number of failures.}

\item{m}{Number of repairs.}

\item{k}{Number of Admin/Logistics Delays (set to NA as default to exclude ALDT).}

\item{mttr}{Assumed mean time to repair.}

\item{mldt}{Assumed mean logistics delay time (set to NA as default to exclude ALDT).}

\item{ao}{Availability to be demonstrated.}

\item{alpha}{Two-sided alpha value.}
}
\value{
The output will be a numeric vector giving the required Up Time,
  assumed Down Time, and Total Time.
}
\description{
\code{ao_keesee_test_duration} determines the test time required to demonstrate
   Availability was met with confidence, assuming a nunmber of failures, repairs,
  and mean time to repair (MTTR) (an optionally Admin/Logistics Delay Time (ALDT),
  Mean ALDT (MLDT), and the number of observed ALDTs).
}
\examples{
ao_keesee_test_duration(n=1, m=1, mttr=1, ao_lcb=.83, alpha = .4)
ao_keesee_test_duration(n=2, m=2, mttr=10.8+40.4, ao_lcb=.83, alpha = .4)
ao_keesee_test_duration(n=2, m=2, mttr=10.8, ao_lcb=0.9585852, alpha = .4)

}
\references{
Keesee, W.R. A Method of Determining a Confidence Interval for Availability. 1965. url{https://apps.dtic.mil/sti/citations/AD0617716}
}
\seealso{
\code{\link{ao_unknown}}, \code{\link{ai_from_ao}},
  \code{\link{ao_keesee}}
}
