% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ao_unknown.R, R/availability.R
\name{ao_unknown}
\alias{ao_unknown}
\title{Solve for one unknown in the Operational Availability formula}
\usage{
ao_unknown(ao = NA, upt = NA, cmt = NA, aldt = NA)

ao_unknown(ao = NA, upt = NA, cmt = NA, aldt = NA)
}
\arguments{
\item{ao}{Operational Availability.}

\item{upt}{Up Time.}

\item{cmt}{Corrective Maintenance Time.}

\item{aldt}{Administrative and Logistics Down Time.}
}
\value{



}
\description{
\code{ao_unknown} solves for the unknown in the Operational Availability
  equation given the other inputs. One and only one of the parameters
  should be passed as \code{NA}.

\code{ao_unknown} solves for the unknown in the Operational Availability
  equation given the other inputs. One and only one of the parameters
  should be passed as \code{NA}.
}
\examples{
# What is the required MTBF to have a 90\% prob of
  # completing a 24 hour mission duration?
ao_unknown(ao = 0.83, upt = 250, cmt = NA, aldt = 40.4)

# What is the required MTBF to have a 90\% prob of
  # completing a 24 hour mission duration?
ao_unknown(ao = 0.83, upt = 250, cmt = NA, aldt = 40.4)

}
\seealso{
\code{\link{ai_from_ao}}, \code{\link{ao_keesee}},
  \code{\link{ao_keesee_test_duration}}

\code{\link{ai_from_ao}}, \code{\link{keese_ao}},
  \code{\link{keese_test_duration}}
}
