% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asm_app.R
\name{asm_app}
\alias{asm_app}
\title{Shiny app for use with MCOTEA Analysis Sychronization Matrix (ASM)}
\usage{
asm_app(filedir, launch.browser = TRUE)
}
\value{
A Shiny Application.
}
\description{
\code{asm_app} provides a Shiny application for importing and viewing data
  from the MCOTEA ASM datFormTable.xlsx and qryProgramsMilestoneInfo.xlsx output.
}
\examples{
# asm_app(
#   filedir = "directory containing datFormProgram.xlsx and qryProgramsMilestoneInfo.xlsx",
#   launch.browser = TRUE
# )

}
\seealso{

}
