% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binom_prst.R
\name{binom_prst}
\alias{binom_prst}
\title{Probability Ratio Sequential Test (PRST)}
\usage{
binom_prst(
  p0 = p0,
  p1 = p1,
  m = m,
  alpha = 0.1,
  beta = 0.1,
  r = NULL,
  t = NULL
)
}
\arguments{
\item{p0}{The minimum acceptable probability of failure.}

\item{p1}{The maximum acceptable probability of failure. This value must be greater than \code{p0}.}

\item{m}{An integer indicating the maximum number of units allowable for test.}

\item{alpha}{The producer's risk of rejecting equipment with probability of failure p < \code{p0}
(the probility of rejecting good equipment). Default = 0.10.}

\item{beta}{The consumer's risk of accepting equipment with probability of failure p > \code{p1}.
(the probility of accepting bad equipment). Default = 0.10.}

\item{r}{(Optional, default as NULL) The cumulative number of failures up to the number of units tested, \code{t}.
If specified it must be of equal length to t.}

\item{t}{(Optional, default as NULL) The cumulative number of units tested.
If specified it must be of equal length to \code{r}.}
}
\value{
The output will return a table indicating when to accept
or reject the equipment based on the number of units tested and
failures observed. The binomial prst plot will be printed, along
with the Operating Characteristic Curve and Average Sample Number plot.
}
\description{
\code{binom_prst} returns parameters defining a PRST for the binomial distribution as developed by Abraham Wald.
}
\examples{
binom_prst(p0 = 0.05, p1 = 0.10, m=300, alpha=0.10, beta=0.10)
binom_prst(p0 = 0.05, p1 = 0.10, m=300, alpha=0.10, beta=0.20)
binom_prst(p0 = 0.05, p1 = 0.10, m=300, alpha=0.05, beta=0.20)

}
\references{
Wald, Abraham. Sequential Analysis. John Wiley & Sons, 1947.
}
\seealso{
\code{\link{exp_prst}}, \code{\link{exp_prst_plot}}
}
