% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cbPalette}
\alias{cbPalette}
\title{A colorblind-friendly palette with grey}
\format{
An object of class \code{character} of length 8.
}
\usage{
cbPalette
}
\description{
A colorblind-friendly palette vector with 8 hexidecimal colors,
the first of which is grey. This comes from: 
\href{http://www.cookbook-r.com/Graphs/Colors_(ggplot2)/}{Cookbook for R}.
}
\seealso{
\code{\link{cbbPalette}}
}
\keyword{datasets}
