% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_beta.R
\name{common_beta}
\alias{common_beta}
\title{Bartlett's Modified Likelihood Ratio Chi-Sq Test for Common Shape}
\usage{
common_beta(t, T, fail.trunc = FALSE)
}
\arguments{
\item{t}{A list of failure time vectors. Each vector should indicate
a different system, i.e. if you have multiple systems each
systems' failure times should be in it's own vector.}

\item{T}{A list of Total Time on Test (TTT) (i.e. test duration) vectors.
The vectors in the list should be of length 1, and each vector should
indicate a different system, i.e. if you have multiple systems each
systems' TTT should be in it's own vector.}

\item{fail.trunc}{Logical indicating if the test was failure terminated.}
}
\value{
The output will be a list with the test statistic, p-value, and
  degrees of freedom.
}
\description{
\code{common_beta} provides a test for a common shape parameter given multiple
  systems/processes assumed to come from a Nonhomogeneous Poisson Process (NHPP)
  following a Power Law Model (Crow-AMSAA model). Small p-values reject null
  hypothesis of common beta (shape paramter). See: 
  \href{https://support.minitab.com/en-us/minitab/18/help-and-how-to/modeling-statistics/reliability/how-to/parametric-growth-curve/methods-and-formulas/tests-for-equal-shapes-or-scales/}{Minitab Methods and formulas for test for equal shapes or scales in Parametric Growth Curve}
  and \href{https://apps.dtic.mil/docs/citations/ADA020296}{AMSAA Technical Report No. 138 Reliability Analysis for Complex, Repairable Systems (L.H. Crow 1975)}.
}
\examples{
data(amsaa)

common_beta(
  t = split(amsaa$Time, amsaa$System),
  T = list(200, 200, 200),
  fail.trunc = FALSE) 

}
\seealso{
\code{\link{power_law_process}}, \code{\link{power_law_mcf}},
  \code{\link{mcf}}, \code{\link{trend_test}}, \code{\link{ttt}}
}
