% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_fixed_duration_tests.R
\name{exp_fixed_duration_tests}
\alias{exp_fixed_duration_tests}
\title{Exponential Fixed Duration Test Plans}
\usage{
exp_fixed_duration_tests(MTBF0 = 500, MTBFA = 1000, alpha = 0.2, beta = 0.2)
}
\arguments{
\item{MTBF0}{The MTBF below which to reject the system with probability alpha.}

\item{MTBFA}{The MTBF above which to accept the system with probability 1-beta.}

\item{alpha}{The type I error rate, producer's risk - the probability we reject
a system with MTBF > MTBFA.}

\item{beta}{The type II error rate, consumer's risk - the probability we accept
a system with MTBF < MTBF0.}
}
\value{
The output is a data.frame containing the allowable number of failures
  at which the system would pass the test (Accept), the number of failures at
  which the system would fail the test (reject), the test duration (Duration),
  and the actual beta value (the probability we accept a system with MTBF = MTBF0)
  given the proposed test.
}
\description{
\code{exp_fixed_duration_tests} See pg. 67 of MIL-HDBK-781A. Fixed duration
  tests for exponentially distributed MTBF with desired alpha and beta.
}
\examples{
  (fd <- exp_fixed_duration_tests(MTBF0 = 500, MTBFA = 1000, alpha = .2, beta = .2))
  theta <- seq(250, 2000, 50)
  prob_pass <- exp_oc(accept = fd$Accept[6], duration = fd$Duration[6], mtbf = theta)

  df <- data.frame(x = c(500,1000), y = c(0.20, 1-fd$beta[6]), label = c("alpha", "1-beta"))

  ggplot2::ggplot(data.frame(prob_pass = prob_pass, theta = theta)) +
    ggplot2::geom_point(ggplot2::aes(x = theta, y = prob_pass)) +
    ggplot2::scale_y_continuous(limits = c(0,1), breaks = seq(0,1,.2)) +
    ggplot2::geom_point(
      data = df,
      mapping = ggplot2::aes(x, y), colour = "red"
    ) +
    ggplot2::geom_text(
      data = df,
      mapping = ggplot2::aes(x, y, label = label), parse = TRUE, colour = "red", hjust = c(-1,1.2)
    )

}
\references{
Mil-Hdbk-781A
}
\seealso{
\code{\link{exp_mtbf_req}}, \code{\link{exp_reliability_req}},
  \code{\link{exp_test_duration}}, \code{\link{exp_mean_lcb}},
  \code{\link{exp_mean_ci}}, \code{\link{exp_oc}}
}
