% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_mean_ci.R, R/reliability.R
\name{exp_mean_ci}
\alias{exp_mean_ci}
\title{Calculates 2-sided CI for an exponential mean}
\usage{
exp_mean_ci(n, duration, conf = 0.8)

exp_mean_ci(n, duration, conf = 0.8)
}
\arguments{
\item{n}{A numeric vector indicating the observed number of failures.}

\item{duration}{A numeric vector indicating the tested duration
(hours, miles, rounds, etc.).}

\item{conf}{The desired level of confidence (values within 0.0:1.00
default is set to 0.80).}
}
\value{
The output will be a numeric vector, indicating the LCB and UCB of
  the exponential mean with the given level of confidence. The units
  will be the same as the units of the supplied \code{duration}.

The output will be a numeric vector, indicating the LCB and UCB of
  the exponential mean with the given level of confidence. The units
  will be the same as the units of the supplied \code{duration}.
}
\description{
\code{exp_mean_ci} calculates the 2-sided CI
  for an exponential mean given a test duration (hours, miles,
  rounds, etc.), number of failures, confidence level, and assuming
  the test is time terminated.

\code{exp_mean_ci} calculates the 2-sided CI
  for an exponential mean given a test duration (hours, miles,
  rounds, etc.), number of failures, confidence level, and assuming
  the test is time terminated.
}
\examples{
# What is the 80\% 2-sided CI for the MTBF (assuming the times between 
  # failure are exponentially distributed), given a test
  # time of 367 hours, and 0 failures.
exp_mean_ci(n = 0, duration = 367, conf = 0.80)

# What is the 80\% 2-sided CI for the MTBF (assuming the times between 
  # failure are exponentially distributed), given a test
  # time of 920 hours, and 7 failures.
exp_mean_ci(n = 7, duration = 920, conf = 0.80)

# What is the 80\% 2-sided CI for the MTBF (assuming the times between 
  # failure are exponentially distributed), given a test
  # time of 367 hours, and 0 failures.
exp_mean_ci(n = 0, duration = 367, conf = 0.80)

# What is the 80\% 2-sided CI for the MTBF (assuming the times between 
  # failure are exponentially distributed), given a test
  # time of 920 hours, and 7 failures.
exp_mean_ci(n = 7, duration = 920, conf = 0.80)

}
\seealso{
\code{\link{mtbf_req}}, \code{\link{reliability_req}},
  \code{\link{test_duration}}, \code{\link{test_demo}},
  \code{\link{exp_equal_mtbf}}

\code{\link{mtbf_req}}, \code{\link{reliability_req}},
  \code{\link{test_duration}}, \code{\link{test_demo}},
  \code{\link{exp_equal_mtbf}}
}
