% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_prst.R
\name{exp_prst}
\alias{exp_prst}
\title{Probability Ratio Sequential Test (PRST)}
\usage{
exp_prst(mtbf0, mtbfa, alpha = 0.1, beta = 0.1)
}
\arguments{
\item{mtbf0}{The minimum acceptable MTBF.}

\item{alpha}{The producer's risk of rejecting equipment with MTBF > MTBFA
(the probility of rejecting good equipment). Default = 0.10.}

\item{beta}{The consumer's risk of accepting equipment with MTBF < MTBF0.
(the probility of accepting bad equipment). Default = 0.10.}

\item{mtbfA}{A value of MTBF greater than MTBF0 at which we want to
ensure we don't reject the system.}
}
\value{
The output will return \code{r} = the number of failures at which
  the test terminated, \code{T} = the time at which the test is terminated,
  \code{a} = y-intercept for the accept time, \code{b} = slope of the accept
  and reject lines, \code{c} = y-intercept of the reject line, \code{df} =
  a \code{data.frame} giving the number of failures and the associated time
  times at which the test is terminated and the equipment accepted or rejected,
  as well as some text which outputs the form of the accept/reject lines, and
  the accept/reject times.
}
\description{
\code{exp_prst} returns parameters defining a PRST for the exponential distribution as developed by Abraham Wald.
}
\examples{
exp_prst(mtbf0=100, mtbfa=200, alpha=.1, beta=.1)
exp_prst(mtbf0=100, mtbfa=200, alpha=.2, beta=.1)

}
\references{
Mil-Hdbk-781A

Wald, Abraham. Sequential Analysis. John Wiley & Sons, 1947.

Brazovsky, Igor. Reliability Theory and Practice. Prentice Hall, 1961.
}
\seealso{
\code{\link{exp_reliability_req}}, \code{\link{exp_test_duration}},
  \code{\link{exp_test_demo}}, \code{\link{exp_prst_plot}}
}
