% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_residual_plots.R
\name{gg_residual_plots}
\alias{gg_residual_plots}
\title{Creates residual plots in ggplot2}
\usage{
gg_residual_plots(x, ...)
}
\arguments{
\item{x}{A data.frame object (such as that returned from
\code{\link{df_from_model}}) that contains two columns, the first
of which is the residuals, and the second is fitted values.}
}
\value{
Residual plot(s) useful for checking model assumptions.
}
\description{
\code{gg_residual_plots} creates a four-pack of plots similar to the
residual plots returned by Minitab. Plots include (1) a Normal QQ plot of
the residuals, (2) a residuals vs fitted values plot, (3) a histogram 
of the residuals, and (4) a plot of residuals vs order.
The 4th plot (residuals vs order),
assumes the data.frame supplied is already in order.
}
\examples{
data(npk)
npk.aov <- aov(yield ~ block + N*P*K, npk)
summary(npk.aov)
npk.aov1 <- aov(yield ~ block + N + K, data = npk)
summary(npk.aov1)
summary.lm(npk.aov1)
model.tables(npk.aov1, type = "means", se = TRUE)
plot(npk.aov1)

npk.table <- df_from_model(npk.aov1, type = "rstandard")
gg_residual_plots(npk.table)

}
\seealso{
\link[stats]{plot.lm}, \link{residual_plots}, \link{df_from_model}
}
