% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/availability.R
\name{keese_test_duration}
\alias{keese_test_duration}
\title{Determines the test time required to demonstrate Availability was met with Confidence}
\usage{
keese_test_duration(n, m, mttr, ao, alpha = 0.4)
}
\arguments{
\item{n}{Number of failures.}

\item{m}{Number of repairs.}

\item{mttr}{Assumed mean time to repair.}

\item{ao}{Availability to be demonstrated.}

\item{alpha}{Two-sided alpha value.}
}
\value{
The output will be a numeric vector giving the required Up Time,
  assumed Down Time, and Total Time.
}
\description{
\code{keese_test_duration} determines the test time required to demonstrate
   Availability was met with confidence, assuming a nunmber of failures, repairs,
  and mean time to repair (MTTR).
}
\examples{
keese_test_duration(n=1, m=1, mttr=1, ao_lcb=.83, alpha = .4)
keese_test_duration(n=2, m=2, mttr=10.8+40.4, ao_lcb=.83, alpha = .4)
keese_test_duration(n=2, m=2, mttr=10.8, ao_lcb=0.9585852, alpha = .4)

}
\seealso{
\code{\link{ao_unknown}}, \code{\link{ai_from_ao}},
  \code{\link{keese_ao}}
}
