% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repairable_systems_analysis.R
\name{mcf}
\alias{mcf}
\title{Nonparametric Estimate for the Mean Cumulative Function}
\usage{
mcf(t, by = NULL)
}
\arguments{
\item{t}{A list of failure time vectors. Each vector should indicate
a different system, i.e. if you have multiple systems each
systems' failure times should be in it's own vector.}

\item{by}{If providing a list of length > 1 this can be
a vector that defines a name for each element of the list
so as to return by system mcf estimates.}
}
\value{
The output will be a \code{\link[base]{data.frame}} containing,
  the ordered failure times ("t") and corresponding nonparametric
   mcf estimate ("mcf").
}
\description{
\code{mcf} calculates the sample nonparametric estimate of the mean
  cumulative function (expected number of failures at time t).
}
\examples{
data(amsaa)

# Three systems failure times.
mcf(t = split(amsaa$Time, amsaa$System))

}
\seealso{
\code{\link{power_law_process}}, \code{\link{power_law_mcf}},
  \code{\link{trend_test}}, \code{\link{ttt}}, \code{\link{common_beta}}
}
