% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reliability.R
\name{mtbf_req}
\alias{mtbf_req}
\title{Calculates the Mean Time Between Failure (MTBF) Requirement}
\usage{
mtbf_req(r, md)
}
\arguments{
\item{r}{A numeric Reliability requirement (0.00:1.00).}

\item{md}{A numeric mission duration. This could be time, miles, rounds, etc.}
}
\value{
The output will be a numeric vector with units the same
  as the mission duration.
}
\description{
\code{mtbf_req} calculates the Mean Time Between Failure
  (MTBF) requirement based on a defined Reliability requirement,
  and mission duration. (Instead of Time, Miles, Rounds, etc. may be used.)
  This assumes the times (miles, rounds, etc.) between failure
  are exponentially distributed.
}
\examples{
# What is the required MTBF to have a 90\% prob of
  # completing a 24 hour mission duration?
mtbf_req(r = .9, md = 24)

}
\seealso{
\code{\link{reliability_req}}, \code{\link{test_duration}},
  \code{\link{exp_mean_lcb}}
}
