% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prst_plot.R
\name{prst_plot}
\alias{prst_plot}
\title{Probability Ratio Sequential Test (PRST) Plot}
\usage{
prst_plot(
  r = NULL,
  t = NULL,
  mtbf0 = 100,
  mtbfa = 200,
  alpha = 0.1,
  beta = 0.1
)
}
\arguments{
\item{r}{(Optional, default as NULL) The cumulative number of failures up to time t.
If specified it must be of equal length to t.}

\item{t}{(Optional, default as NULL) The cumulative (operational) time.
If specified it must be of equal length to r.}

\item{mtbf0}{The minimum acceptable MTBF.}

\item{mtbfa}{A value of MTBF greater than MTBF0 at which we want to
ensure we don't reject the system.}

\item{alpha}{The producer's risk of rejecting equipment with MTBF > MTBFA
(the probility of rejecting good equipment). Default = 0.10.}

\item{beta}{The consumer's risk of accepting equipment with MTBF < MTBF0.
(the probility of accepting bad equipment). Default = 0.10.}
}
\value{
The output will return \code{r} = the number of failures at which
  the test terminated, \code{T} = the time at which the test is terminated,
  \code{a} = y-intercept for the accept time, \code{b} = slope of the accept
  and reject lines, \code{c} = y-intercept of the reject line, \code{df} =
  a \code{data.frame} giving the number of failures and the associated time
  times at which the test is terminated and the equipment accepted or rejected,
  as well as some text which outputs the form of the accept/reject lines, and
  the accept/reject times. A plot will also be returned plotting the accept/reject
  lines, truncation lines, and the observed data as step function if supplied by
  parameters \code{r} and \code{t}.
}
\description{
\code{prst_plot} returns parameters defining a PRST (as developed by Abraham Wald),
  and creates the plot of the results and defined parameters.
}
\examples{
prst_plot(r=NULL, t=NULL, mtbf0=100, mtbfa=200, alpha=.1, beta=.1)
prst_plot(
  r = c(0,1,2,3,4,5,6,6,14),
  t = c(0,100,150,200,250,275,300,1500,2060),
  mtbf0=100, mtbfa=200, alpha=.1, beta=.1)

}
\references{
Mil-Hdbk-781A

Wald, Abraham. Sequential Analysis. John Wiley & Sons, 1947.

Brazovsky, Igor. Reliability Theory and Practice. Prentice Hall, 1961.
}
\seealso{
\code{\link{reliability_req}}, \code{\link{test_duration}},
  \code{\link{test_demo}}, \code{\link{prst}}
}
