% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rel_conf.R, R/reliability.R
\name{rel_conf}
\alias{rel_conf}
\title{Calcualtes the confidence that we've met the reliability requirement}
\usage{
rel_conf(mtbf_req, duration, r)

rel_conf(mtbf_req, duration, r)
}
\arguments{
\item{mtbf_req}{A numeric vector indicating the MTBF requirement.}

\item{duration}{A numeric vector indicating the tested duration
(hours, miles, rounds, etc.).}

\item{r}{A numeric vector indicating the observed number of failures.}
}
\value{
The output will be a numeric vector, indicating the confidence
  that the MTBF requirement was met.

The output will be a numeric vector, indicating the confidence
  that the MTBF requirement was met.
}
\description{
\code{rel_conf} calculates the confidence that we've met the reliability
  requirement (the true MTBF exceeds the MTBF requirement), assuming that
  the times between falilure are exponentially distributed.

\code{rel_conf} calculates the confidence that we've met the reliability
  requirement.
}
\examples{
# How much confidence do we have that the MTBF exceeds 250 hours, given
  # that we tested for 500 hours and observed 0 failures
rel_conf(250, 500, 0)

# How much confidence do we have that the MTBF exceeds 250 hours, given
  # that we tested for 500 hours and observed 0 failures
rel_conf(250, 500, 0)

}
\seealso{
\code{\link{mtbf_req}}, \code{\link{reliability_req}},
  \code{\link{test_duration}}, \code{\link{test_demo}},
  \code{\link{exp_equal_mtbf}}, \code{\link{exp_mean_ci}},
  \code{\link{exp_mean_lcb}}

\code{\link{mtbf_req}}, \code{\link{reliability_req}},
  \code{\link{test_duration}}, \code{\link{test_demo}},
  \code{\link{exp_equal_mtbf}}, \code{\link{exp_mean_ci}},
  \code{\link{exp_mean_lcb}}
}
