% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reliability_req.R
\name{reliability_req}
\alias{reliability_req}
\title{Calculates the Reliability Requirement}
\usage{
reliability_req(mtbf, md)
}
\arguments{
\item{mtbf}{A numeric MTBF requirement. This could be time, miles, rounds,
etc. Units should match the \code{md} parameter.}

\item{md}{A numeric mission duration. This could be time, miles, rounds,
etc. Units should match the \code{mtbf} parameter.}
}
\value{
The output will be a numeric vector defining the Reliability
  requirement, (the probability that no failure is observed during
  the given mission duration).
}
\description{
\code{reliability_req} calculates the Reliability requirement based on a defined
  Mean Time Between Failure (MTBF) (really we use Mean Time Between
  Operational Mission Failure (MTBOMF)) requirement, and mission duration.
  (Instead of Time, Miles, Rounds, etc. may be used.)
  This assumes the times (miles, rounds, etc.) between failure
  are exponentially distributed.
}
\examples{
# What is the required Reliability if the MTBF is 228 hours and
  # a mission duration of 24 hours?
reliability_req(mtbf = 228, md = 24)

}
\seealso{
\code{\link{mtbf_req}}, \code{\link{test_duration}},
  \code{\link{exp_mean_lcb}}, \code{\link{test_demo}},
  \code{\link{exp_equal_mtbf}}
}
