% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repairable_systems_analysis.R
\name{rocof}
\alias{rocof}
\title{Nonparametric Estimate for the Rate of Occurrence of Failures (ROCOF)}
\usage{
rocof(t, by = NULL)
}
\arguments{
\item{t}{A list of failure time vectors. Each vector should indicate
a different system, i.e. if you have multiple systems each
systems' failure times should be in it's own vector.}

\item{by}{If providing a list of length > 1 this can be
a vector that defines a name for each element of the list
so as to return by system rocof and mtbf estimates.}
}
\value{
The output will be a \code{\link[base]{data.frame}} containing,
  the 'by' variable (which specifies the system or process the failuresm
  are attributable to, if supplied in the list name), the failure times ("t"),
  and the corresponding nonparametric estimates for the instantaneous
  rocof and mtbf.
}
\description{
\code{rocof} calculates the sample nonparametric estimate of the
  intensity function (i.e. rate of occurrence of failures (ROCOF), or failure rate)
  at time t. The inverse of this would be the sample nonparametric
  mean time between failure at time t.
}
\examples{
data(amsaa)

# Three systems failure times.
rocof(t = split(amsaa$Time, amsaa$System))

}
\seealso{
\code{\link{power_law_process}}, \code{\link{power_law_mcf}},
  \code{\link{trend_test}}, \code{\link{ttt}}, \code{\link{common_beta}}
}
