% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sus_score.R
\name{sus_score}
\alias{sus_score}
\title{Calcultes a SUS Score}
\usage{
sus_score(x)
}
\arguments{
\item{x}{A vector of SUS responses for a single record/observation.
the response to a different SUS question.}
}
\value{
Returns a vector of SUS scores.
}
\description{
\code{sus_score} calcultes SUS scores given responses arranded
  in a data.frame or matrix.
}
\examples{
sus_score(x = rep(c(1,5), times = 5))
sus_score(x = rep(c(5,1), times = 5))

}
\seealso{
\code{\link{sus_score_df}}, \code{\link{sus_to_percentile}},
  \code{\link{gg_suaro}}, \code{\link{gg_sus}}
}
