% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sus_score_df.R
\name{sus_score_df}
\alias{sus_score_df}
\title{Calculte SUS Scores}
\usage{
sus_score_df(df)
}
\arguments{
\item{df}{A data.frame or matrix of SUS scores. Each row should be a
different record/observation. There should be 10 columns,
each representing the response to a different SUS question.}
}
\value{
Returns a vector of SUS scores.
}
\description{
\code{sus_score_df} calcultes SUS scores given responses arranded
  in a data.frame or matrix.
}
\examples{
df <- data.frame(
  x1 = c(1,5), x2 = c(5,1), x3 = c(1,5), x4 = c(5,1),
  x5 = c(1,5), x6 = c(5,1), x7 = c(1,5), x8 = c(5,1),
  x9 = c(1,5), x10 = c(5,1)
)
sus_score_df(df)
rm(list = "df")

}
\seealso{
\code{\link{sus_score}}, \code{\link{sus_to_percentile}},
  \code{\link{sus_suaro}}, \code{\link{gg_sus}}
}
