% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SUS.R
\name{sus_suaro}
\alias{sus_suaro}
\title{Creates a SUS plot with acceptability criteria indicated}
\usage{
sus_suaro()
}
\arguments{
\item{...}{Optional parameters which can be supplied to
\code{\link[ggplot2:geom_path]{geom_line}}.}
}
\value{
Returns a base plot of SUS with acceptability criteria.
Line segments and text can be added on top of the base plot
}
\description{
\code{sus_suaro} creates a SUS plot as shown in Suaro's book.
Line segments and text can be added on top of the base plot.
}
\examples{
#setwd("C:/Pictures")
# ppi <- 300
# png("SUS.png", height = 2.5*ppi, width = 8*ppi, res = ppi)
sus_suaro()
segments(x0 = 63.2, y0 = 0, x1 = 63.2, y1 = .6,
  col = "red", lty = 2)
text(x = 63.2, y = .63, adj = c(.5,0),
  label = "IOT = 63.2", col = "red", cex = .8)
# dev.off()
}
\seealso{
\code{\link{sus_to_percentile}}, \code{\link{gg_sus}}
}
