% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SUS.R
\name{sus_to_percentile}
\alias{sus_to_percentile}
\title{Converts a System Usability Scale score to a corresponding percentile rank}
\usage{
sus_to_percentile(x)
}
\arguments{
\item{x}{A valid SUS score ranging within the range 0:100.}
}
\value{
Returns a \code{\link[base]{data.frame}} of x- (SUS scores),
  and y-values (percentiles).
}
\description{
\code{sus_to_percentile} converts a valid SUS score (0:100), into
  the corresponding percentile rank (0.00:1.00).
}
\examples{
sus_to_percentile(c(68, 90))
}
\seealso{
\code{\link{gg_sus}}, \code{\link{sus_suaro}}
}
