% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reliability.R
\name{test_demo}
\alias{test_demo}
\title{Reliability Test Length from Null and Alternative Hypotheses, alpha, and beta}
\usage{
test_demo(mtbf0, mtbf1, alpha = 0.2, beta = 0.2)
}
\arguments{
\item{mtbf0}{A numeric value indicating the null hypothesis for MTBF/MTTR.}

\item{mtbf1}{A numeric value indicating the alternative hypothesis for MTBF/MTTR.
(hours, miles, rounds, etc.).}

\item{alpha}{The desired level of significance (type 1 error rate) (values
within 0.0:1.00, default is set to 0.20).}

\item{beta}{The maximum allowable type II error rate (values
within 0.0:1.00, default is set to 0.20).}
}
\value{
The output is a list supplying required test length (T),
  allowable number of failures (r), and true type II error rate (Beta)
  (type II error rate = probability the null hypothesis is not rejected
  when it is false).
}
\description{
\code{test_demo} calculates the required test length, and allowable number
  of failures for a reliability demonstration test, given null and alternative
  hypotheses for MTTF/MTBF, alpha and beta error rates, and assuming the
  times are exponentially distributed. Hypothesis should be of the form,
  h0: MTBF <= X; h1: MTBF > X+Y, where X and Y are both positive values.
}
\examples{
# What is the required test length and allowable number of failures,  
  # to demonstrate MTBF > 180, with 80\% power and confidence, and allowing 
  # one failure (and assuming the times between failure are
  # exponentially distributed)?
test_demo(mtbf0 = 180, mtbf1 = 300, alpha = 0.2, beta = 0.2)

}
\seealso{
\code{\link{mtbf_req}}, \code{\link{reliability_req}},
  \code{\link{test_duration}}, \code{\link{exp_mean_lcb}},
  \code{\link{rel_power}}, \code{\link{exp_equal_mtbf}}
}
