% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repairable_systems_analysis.R
\name{trend_test}
\alias{trend_test}
\title{Trend Tests for Repairable Systems Analysis}
\usage{
trend_test(t, T, fail.trunc = TRUE)
}
\arguments{
\item{t}{A list of failure time vectors. Each vector should indicate
a different system, i.e. if you have multiple systems each
systems' failure times should be in it's own vector.}

\item{T}{A list of Total Time on Test (TTT) (i.e. test duration) vectors.
The vectors in the list should be of length 1, and each vector should
indicate a different system, i.e. if you have multiple systems each
systems' TTT should be in it's own vector.}

\item{fail.trunc}{Logical indicating if the test was failure terminated.}
}
\value{
The output will be a list identifying the tests, test statistics,
  degrees of freedom (where applicable), and p-values.
}
\description{
\code{trend_test} tests to distinguish
between "no trend" and trends in Poisson Processes.
a trend following the Nonhomogeneous. \cr \cr
Laplace Centroid Test: Optimal for distinguishing between "no trend"
and a trend following the Nonhomogeneous Poisson Process (NHPP)
Exponential Law model. \cr \cr
Military Handbook Test: From Mil-HDBK-189, is optimal for
distinguishing between "no trend" and a trend following
the NHPP Power Law or Duane model. \cr \cr See:
\href{https://www.itl.nist.gov/div898/handbook/apr/section2/apr234.htm}{NIST Trend Tests}
}
\examples{
data(amsaa)

# Three systems all time truncated at 200 hours
trend_test(
  t = split(amsaa$Time, amsaa$System),
  T = list(200,200,200),
  fail.trunc = FALSE)

# Three systems all failure truncated
trend_test(
  t = split(amsaa$Time, amsaa$System),
  T = list(197.2,190.8,195.8),
  fail.trunc = TRUE)

#' # One system, time truncated
trend_test(
  t = list(subset(amsaa$Time, amsaa$System == "S1")),
  T = list(200),
  fail.trunc = FALSE)

# One system, failure truncated
trend_test(
  t = list(subset(amsaa$Time, amsaa$System == "S1")),
  T = list(197.2),
  fail.trunc = TRUE)

}
\seealso{
\code{\link{power_law_process}}, \code{\link{power_law_mcf}},
  \code{\link{mcf}}, \code{\link{ttt}}, \code{\link{common_beta}}
}
