% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weibull_oc.R
\name{weibull_oc}
\alias{weibull_oc}
\title{Weibull Operating Characteristic (OC) Curves}
\usage{
weibull_oc(accept, ss, duration, shape, scale = NULL, md, R_m = NULL, plot = T)
}
\arguments{
\item{accept}{The allowable number of failures.}

\item{duration}{The test duration.}

\item{shape}{The Weibull shape parameter.}

\item{scale}{The Weibull scale parameter (could be left as NULL if R_m is supplied).
This is \code{NULL} by default.}

\item{md}{The required mission duration for the reliability requirement.}

\item{R_m}{The reliability requirement for the given mission duration.}

\item{s}{The number of test units.}
}
\value{
Returns a data.frame with the probability of passing the test of a given
  duration with an allowable number of failures, given the true MTBF and R_m.
}
\description{
\code{weibull_oc} computes OC curves for Weibull reliability tests
  (probability of passing test as a function of true prob or shape & scale)
}
\examples{
shape <- 2
scale_vec <- seq(4000, 6000, by = 2000)
mtbf_vec <- weibull_mean(shape, scale_vec)
weibull_oc(accept = 1, ss = 32, duration = 1500, shape = shape,
  scale = scale_vec, md = 2000, R_m = NULL, plot = F)

weibull_oc(accept = 1, ss = 32, duration = 1500, shape = shape,
  scale = NULL, md = 2000, R_m = c(.7788008, 0.8948393), plot = F)


scale_vec <- seq(4000, 30000, by = 50)
scale <- weibull_scale(2000, R_m = .8, shape = shape)
mtbf <- weibull_mean(shape, scale)
p1 <- weibull_oc(accept = 1, ss = 32, duration = 1500, shape = shape,
  scale = scale_vec, md = 2000, R_m = NULL, plot = T)
p1$p + 
  scale_y_continuous(limits = c(0,1), breaks = seq(0,1,.2)) +
  scale_x_continuous(limits = c(mtbf,10000), breaks = seq(3000,10000,1000))

}
