% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weibull_rel_demo.R
\name{weibull_rel_demo}
\alias{weibull_rel_demo}
\title{Calculate the Required Test Duration or Sample Size}
\usage{
weibull_rel_demo(t, ss, r, alpha, shape, md, R_m, lower, upper, ...)
}
\arguments{
\item{t}{The test duration. Set to \code{NULL} to solve for test duration. Default is \code{NULL}.}

\item{ss}{The sample size (number of units) to be used in the test.
Set to \code{NULL} to solve for sample size. Default is 1. If set it must be > \code{r}.}

\item{r}{The allowable number of failures in the test. Default is 0.}

\item{alpha}{The alpha value (Type I Error Rate) to be used in the test. Default is 0.20.}

\item{shape}{The Weibull shape parameter. Default is 1 corresponding to an exponential distribution.
Shape > 1 indicates increasing failure rate, Shape < 1 indicates decreasing failure rate.}

\item{md}{The mission duration requirement.}

\item{R_m}{The reliability requirement for the specified mission duration.}

\item{lower}{The lower bound in the search range (the lower of where to look for t or ss).
Required for the \code{\link[stats]{uniroot}} or \code{\link[ssanv]{uniroot.integer}} functions.}

\item{upper}{The upper bound in the search range (the upper of where to look for t or ss).
Required for the \code{\link[stats]{uniroot}} or \code{\link[ssanv]{uniroot.integer}} functions.}

\item{...}{Additional options parameter to be passed to the \code{\link[stats]{uniroot}} or
\code{\link[ssanv]{uniroot.integer}}functions.}
}
\value{
The output will be from the \code{\link[stats]{uniroot}} or \code{\link[ssanv]{uniroot.integer}} functions.
  The root element of the return (as in \code{weibull_rel_demo(...)$root}) provides the solution.
}
\description{
\code{weibull_rel_demo} calculates the required test duration or sample size
  (number of units to be tested) based on the alpha value, Weibull shape parameter,
  mission duration requirement and corresponding reliability requirement. Leave one of
  \code{t} or \code{ss} set to \code{NULL} to solve for it.
}
\examples{
# What is the required sample size given 1 allowable failure
  # in 1500 hours to demonstrate with 90\% confidence that
  # a system has 0.80 probability of surviving to 2000 hours
  # without failure?
weibull_rel_demo(t = 1500, ss = NULL, r = 1, alpha = .1,
  shape = 2, md = 2000, R_m = .8, lower=1, upper = 100)

# What is the required test duration (per system), if I have 10 systems,
  # allow for a total of 1 failure, in order to demonstrate with 90\%
  # confidence that a system has 0.80 probability of surviving to 2000
  # hours without failure?
weibull_rel_demo(t = NULL, ss = 10, r = 1, alpha = .1,
  shape = 2, md = 2000, R_m = .8, lower=1, upper = 10000)

# These are equivalent since shape = 1 reduces to exponential
# However, if the number of allowable failures (r) increases, the ss must increase and
  # the return from the exp_test_duration function will not match
weibull_rel_demo(t = NULL, ss = 1, r = 0, alpha = .1,
  shape = 1, md = 2000, R_m = .8, lower=1, upper = 100000)
mtbf_req(.8, 2000)
test_duration(r = 0, mtbf = 8962.84, conf = .9)

}
\references{

}
\seealso{
\code{\link{weibull_scale}}, \code{\link{weibull_mean}}
}
