% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asm_ag_app.R
\name{asm_ag_app}
\alias{asm_ag_app}
\title{Shiny app for use with MCOTEA Analysis Sychronization Matrix (ASM)}
\usage{
asm_ag_app(filedir, launch.browser = TRUE)
}
\value{
A Shiny Application.
}
\description{
\code{asm_ag_app} provides a Shiny application for importing and viewing data
  from the MCOTEA ASM datFormTable.xlsx output.
}
\examples{
# asm_ag_app(
#   filedir = "directory containing datFormProgram.xlsx"
#   launch.browser = TRUE
# )

}
