% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbbPalette.R, R/data.R
\docType{data}
\name{cbbPalette}
\alias{cbbPalette}
\title{A colorblind-friendly palette with black}
\format{
An object of class \code{character} of length 8.

An object of class \code{character} of length 8.
}
\usage{
cbbPalette

cbbPalette
}
\description{
A colorblind-friendly palette vector with 8 hexidecimal colors,
the first of which is black. This comes from: 
\href{http://www.cookbook-r.com/Graphs/Colors_(ggplot2)/}{Cookbook for R}.

A colorblind-friendly palette vector with 8 hexidecimal colors,
the first of which is black. This comes from: 
\href{http://www.cookbook-r.com/Graphs/Colors_(ggplot2)/}{Cookbook for R}.
}
\seealso{
\code{\link{cbPalette}}

\code{\link{cbPalette}}
}
\keyword{datasets}
