% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_mean_lcb.R
\name{exp_mean_lcb}
\alias{exp_mean_lcb}
\title{Calculates the lower confidence bound for an exponential mean}
\usage{
exp_mean_lcb(r, duration, alpha = 0.2)
}
\arguments{
\item{r}{A numeric vector indicating the observed number of failures.}

\item{duration}{A numeric vector indicating the tested duration
(hours, miles, rounds, etc.).}

\item{alpha}{The allowable type I error rate (1-confidence).
Values must be within 0.0:1.00 default is set to 0.20.}
}
\value{
The output will be a numeric vector, indicating the LCB of
  the exponential mean with the given level of confidence. The units
  will be the same as the units of the supplied \code{duration}.
}
\description{
\code{exp_mean_lcb} calculates the lower confidence bound (LCB)
  for an exponential mean given a test duration (hours, miles,
  rounds, etc.), number of failures, confidence level, and assuming
  the test is time terminated.
}
\examples{
# What is the 80\% LCB for the MTBF (assuming the times between 
  # failure are exponentially distributed), given a test
  # time of 367 hours, and 0 failures.
exp_mean_lcb(r = 0, duration = 367, alpha = 0.20)

}
\seealso{
\code{\link{exp_mtbf_req}}, \code{\link{exp_reliability_req}},
  \code{\link{exp_test_duration}}, \code{\link{exp_equal_mtbf}}
  \code{\link{exp_fixed_duration_tests}}
}
